fix supporting for workforce pool and split module due to lines limit
