+++ lib/googleauth/external_account/aws_credentials.rb
@@ -14,6 +14,7 @@
 
 require "time"
 require "googleauth/external_account/base_credentials"
+require "googleauth/external_account/external_account_utils"
 
 module Google
   # Module Auth provides classes that provide Google-specific authorization used to access Google APIs.
@@ -27,6 +28,7 @@ module Google
         IMDSV2_TOKEN_EXPIRATION_IN_SECONDS = 300
 
         include Google::Auth::ExternalAccount::BaseCredentials
+        include Google::Auth::ExternalAccount::ExternalAccountUtiles
         extend CredentialsLoader
 
         # Will always be nil, but method still gets used.
