+++ lib/googleauth/external_account/base_credentials.rb
@@ -72,54 +72,6 @@ module Google
           notify_refresh_listeners
         end
 
-        ##
-        # Retrieves the project ID corresponding to the workload identity or workforce pool.
-        # For workforce pool credentials, it returns the project ID corresponding to the workforce_pool_user_project.
-        # When not determinable, None is returned.
-        #
-        # The resource may not have permission (resourcemanager.projects.get) to
-        # call this API or the required scopes may not be selected:
-        # https://cloud.google.com/resource-manager/reference/rest/v1/projects/get#authorization-scopes
-        #
-        # @return [string,nil]
-        #     The project ID corresponding to the workload identity pool or workforce pool if determinable.
-        #
-        def project_id
-          return @project_id unless @project_id.nil?
-          project_number = self.project_number || @workforce_pool_user_project
-
-          # if we missing either project number or scope, we won't retrieve project_id
-          return nil if project_number.nil? || @scope.nil?
-
-          url = "#{CLOUD_RESOURCE_MANAGER}#{project_number}"
-
-          response = connection.get url do |req|
-            req.headers["Authorization"] = "Bearer #{@access_token}"
-            req.headers["Content-Type"] = "application/json"
-          end
-
-          if response.status == 200
-            response_data = MultiJson.load response.body, symbolize_names: true
-            @project_id = response_data[:projectId]
-          end
-
-          @project_id
-        end
-
-        ##
-        # Retrieve the project number corresponding to workload identity pool
-        # STS audience pattern:
-        #     `//iam.googleapis.com/projects/$PROJECT_NUMBER/locations/...`
-        #
-        # @return [string, nil]
-        #
-        def project_number
-          segments = @audience.split "/"
-          idx = segments.index "projects"
-          return nil if idx.nil? || idx + 1 == segments.size
-          segments[idx + 1]
-        end
-
         # Retrieves the subject token using the credential_source object.
         # @return [string]
         #     The retrieved subject token.
@@ -160,27 +112,19 @@ module Google
           )
         end
 
-        def normalize_timestamp time
-          case time
-          when NilClass
-            nil
-          when Time
-            time
-          when String
-            Time.parse time
-          else
-            raise "Invalid time value #{time}"
-          end
-        end
-
         def exchange_token
+          additional_options = nil
+          if @client_id.nil? && @workforce_pool_user_project
+            additional_options = {:userProject => @workforce_pool_user_project}
+          end
           @sts_client.exchange_token(
             audience: @audience,
             grant_type: STS_GRANT_TYPE,
             subject_token: retrieve_subject_token!,
             subject_token_type: @subject_token_type,
             scopes: @service_account_impersonation_url ? IAM_SCOPE : @scope,
-            requested_token_type: STS_REQUESTED_TOKEN_TYPE
+            requested_token_type: STS_REQUESTED_TOKEN_TYPE,
+            additional_options: additional_options,
           )
         end
